requirements.txt
jpype1
numpy
scipy
torch